# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/user-media/user/{userId}": {
      "get": {
        "operationId": "UserMediaController_getMediaByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserMediaResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserMedia"
        ]
      }
    },
...
    "/user-tributes/user/{userId}": {
      "get": {
        "operationId": "UserTributeController_getTributesByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserTributeResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserTribute"
        ]
      }
    },
...
    "/user-featured-memorials/user/{userId}": {
      "get": {
        "operationId": "UserFeaturedMemorialController_getFeaturedMemorialsByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadUserFeaturedMemorialResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "summary": "Get all featured memorials for a user",
        "tags": [
          "UserFeaturedMemorial"
        ]
      }
    },
...
    "/userNotificationSettings/user/{userId}": {
      "get": {
        "operationId": "UserNotificationSettingsController_getByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserNotificationSettingResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserNotificationSettings"
        ]
      }
    },
...
    "/user-questionnaires": {
      "post": {
        "operationId": "UserQuestionnaireController_createQuestionnaire",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserQuestionnaireRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserQuestionnaireResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaire"
        ]
      },
      "get": {
        "operationId": "UserQuestionnaireController_getAllQuestionnaires",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserQuestionnaireResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaire"
        ]
      }
    },
    "/user-questionnaires/{id}": {
      "patch": {
        "operationId": "UserQuestionnaireController_updateQuestionnaire",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserQuestionnaireRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserQuestionnaireResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaire"
        ]
      },
      "delete": {
        "operationId": "UserQuestionnaireController_deleteQuestionnaire",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserQuestionnaireResponseDto"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaire"
        ]
      }
    },
    "/user-questionnaires/readById/{id}": {
      "get": {
        "operationId": "UserQuestionnaireController_getQuestionnaireById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserQuestionnaireResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaire"
        ]
      }
    },
    "/user-questionnaire-answers": {
      "post": {
        "operationId": "UserQuestionnaireAnswersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserQuestionnaireAnswerRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserQuestionnaireAnswersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaireAnswers"
        ]
      },
      "get": {
        "operationId": "UserQuestionnaireAnswersController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserQuestionnaireAnswerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaireAnswers"
        ]
      }
    },
    "/user-questionnaire-answers/{id}": {
      "patch": {
        "operationId": "UserQuestionnaireAnswersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserQuestionnaireAnswerRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserQuestionnaireAnswersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaireAnswers"
        ]
      },
      "get": {
        "operationId": "UserQuestionnaireAnswersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserQuestionnaireAnswerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaireAnswers"
        ]
      },
      "delete": {
        "operationId": "UserQuestionnaireAnswersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadUserQuestionnaireAnswerResponseDto"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "UserQuestionnaireAnswers"
        ]
      }
    },
...
  },
  "info": {
    "title": "rememeber-me",
    "description": "API documentation for rememeber-me.env",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserRequestDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "string"
          },
          "callingCode": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "photoId": {
            "type": "string"
          },
          "photoURL": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "street1": {
            "type": "string"
          },
          "street2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          }
        }
      },
      "UserRoles": {
        "type": "string",
        "enum": [
          "admin",
          "superAdmin",
          "user"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "callingCode": {
            "type": "string"
          },
          "defaultRole": {
            "type": "string"
          },
          "street1": {
            "type": "string"
          },
          "street2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postal": {
            "type": "string"
          },
          "photoId": {
            "type": "string"
          },
          "photoURL": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRoles"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "roles",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadUserMediaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-media"
          },
          "mimeType": {
            "type": "string",
            "example": "image/png"
          },
          "fileURL": {
            "type": "string",
            "example": "https://cdn.example.com/uploads/media-file.png"
          },
          "fileId": {
            "type": "string",
            "example": "uuid-of-file"
          },
          "type": {
            "type": "string",
            "example": "IMAGE",
            "description": "Media type (e.g., IMAGE, VIDEO, AUDIO)"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "photoCaption": {
            "type": "string",
            "example": "Vacation Memories"
          },
          "photoDescription": {
            "type": "string",
            "example": "A beautiful sunset at the beach"
          },
          "videoTitle": {
            "type": "string",
            "example": "Grandma’s 80th Birthday Celebration"
          },
          "isMainVideo": {
            "type": "boolean",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 1
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "mimeType",
          "fileURL",
          "type",
          "memorialId",
          "userId",
          "isMainVideo",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "ReadUserTributeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-tribute"
          },
          "authorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "content": {
            "type": "string",
            "example": "You will always be remembered for your kindness and love."
          },
          "type": {
            "type": "string",
            "enum": [
              "memory",
              "tribute",
              "condolence"
            ],
            "example": "tribute"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/image.jpg"
          },
          "imageId": {
            "type": "string"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "status": {
            "type": "string",
            "example": "approved",
            "description": "Moderation status"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "authorName",
          "content",
          "type",
          "imageId",
          "memorialId",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadUserFeaturedMemorialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-featured-memorial"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "memorialId": {
            "type": "string",
            "example": "uuid-of-memorial"
          },
          "isFeature": {
            "type": "boolean",
            "example": true,
            "description": "Whether the memorial is currently featured"
          },
          "memorial": {
            "$ref": "#/components/schemas/ReadMemorialResponseDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "userId",
          "memorialId",
          "isFeature",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ReadUserNotificationSettingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "IsInAppAllowed": {
            "type": "boolean"
          },
          "IsEmailAllowed": {
            "type": "boolean"
          },
          "IsDesktopNotificationAllowed": {
            "type": "boolean"
          },
          "IsUnreadBadgeEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "userId",
          "IsInAppAllowed",
          "IsEmailAllowed",
          "IsDesktopNotificationAllowed",
          "IsUnreadBadgeEnabled",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateUserQuestionnaireRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the questionnaire"
          },
          "questionnaire": {
            "type": "string",
            "description": "Questionnaire text or description"
          },
          "iconURL": {
            "type": "string",
            "description": "Icon URL of the questionnaire"
          },
          "iconId": {
            "type": "string",
            "description": "Icon public ID from Cloudinary or S3"
          }
        },
        "required": [
          "title",
          "questionnaire"
        ]
      },
      "CreateUserQuestionnaireResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateUserQuestionnaireRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the questionnaire"
          },
          "questionnaire": {
            "type": "string",
            "description": "Questionnaire text or description"
          },
          "iconURL": {
            "type": "string",
            "description": "Icon URL of the questionnaire"
          },
          "iconId": {
            "type": "string",
            "description": "Icon public ID from Cloudinary or S3"
          }
        }
      },
      "UpdateUserQuestionnaireResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "ReadUserQuestionnaireResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-questionnaire"
          },
          "title": {
            "type": "string",
            "example": "General Health Questionnaire"
          },
          "questionnaire": {
            "type": "string",
            "example": "What is your medical history?"
          },
          "iconURL": {
            "type": "string",
            "example": "https://cdn.example.com/icons/questionnaire.png"
          },
          "iconId": {
            "type": "string",
            "example": "icon-public-id"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T12:34:56Z"
          }
        },
        "required": [
          "id",
          "title",
          "questionnaire",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserQuestionnaireAnswerRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The ID of the user"
          },
          "questionnaireId": {
            "type": "string",
            "description": "The ID of the questionnaire"
          },
          "answerId": {
            "type": "string",
            "description": "The ID of the selected answer"
          }
        },
        "required": [
          "userId",
          "questionnaireId",
          "answerId"
        ]
      },
      "CreateUserQuestionnaireAnswersResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateUserQuestionnaireAnswerRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The ID of the user"
          },
          "questionnaireId": {
            "type": "string",
            "description": "The ID of the questionnaire"
          },
          "answerId": {
            "type": "string",
            "description": "The ID of the selected answer"
          }
        }
      },
      "UpdateUserQuestionnaireAnswersResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "ReadUserQuestionnaireAnswerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-questionnaire-answer"
          },
          "userId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "questionnaireId": {
            "type": "string",
            "example": "uuid-of-questionnaire"
          },
          "answerId": {
            "type": "string",
            "example": "uuid-of-answer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-09-01T12:34:56Z"
          }
        },
        "required": [
          "id",
          "userId",
          "questionnaireId",
          "answerId",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}